recursos/leccion2/mysql/requirements.txt
# Install the packages from PyPi repository.
PyMySQL==1.0.2
python-dotenv==0.21.0
